# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AvisoStatusCurrentTask = {
  id?: unknown;
  type?: unknown;
} | null;

export interface AvisoStatus {
  status?: string | null;
  balance?: string | null;
  balanceRaw?: number | null;
  totalEarned?: number | null;
  totalTasks?: number | null;
  totalYtDone?: number | null;
  totalYtEarned?: number | null;
  currentTask?: AvisoStatusCurrentTask;
  sleepUntil?: string | null;
  lastUpdated?: string | null;
  log?: string[];
  balanceHistory?: unknown[];
}

export interface AvisoUpdate { [key: string]: unknown }

export interface LoginPhase1Result {
  ok: boolean;
  needOtp: boolean;
  message: string;
}

export interface LoginPhase2Input {
  otp: string;
}

export interface LoginPhase2Result {
  ok: boolean;
  message: string;
}

export type UpdateAvisoStatus200 = {
  ok: boolean;
};