# lib/api-zod/src/generated/types/avisoStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AvisoStatusCurrentTask } from './avisoStatusCurrentTask';

export interface AvisoStatus {
  status?: string | null;
  balance?: string | null;
  balanceRaw?: number | null;
  totalEarned?: number | null;
  totalTasks?: number | null;
  totalYtDone?: number | null;
  totalYtEarned?: number | null;
  currentTask?: AvisoStatusCurrentTask;
  sleepUntil?: string | null;
  lastUpdated?: string | null;
  log?: string[];
  balanceHistory?: unknown[];
}