# lib/api-zod/src/generated/types/loginPhase2Result.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoginPhase2Result {
  ok: boolean;
  message: string;
}